Add Sørensen–Dice and Rand metrics to dataset analysis
